requirements.txt
aniso8601==7.0.0
anyio==3.6.2
anytree==2.8.0
asgiref==3.6.0
Babel==2.12.1
beautifulsoup4==4.12.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
dateparser==1.1.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.5
django-4-jet==1.0.9
django-admin-thumbnails==0.2.6
django-appconf==1.0.5
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-debug-toolbar==3.8.1
django-dotenv==1.4.2
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-extensions==3.2.1
django-filter==22.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-measurement==3.2.4
django-mptt==0.14.0
django-multiupload==0.6.1
django-nine==0.2.7
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-rosetta==0.9.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-extra-fields==3.4.1
drf-yasg==1.21.4
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
flake8==6.0.0
flake8-typing-imports==1.14.0
googletrans==3.0.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.21.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
jet-bridge-base==1.6.3
Jinja2==3.1.2
lark-parser==0.11.0
lib-rql==1.1.7
MarkupPy==1.14
MarkupSafe==2.1.2
mccabe==0.7.0
measurement==3.2.2
mpmath==1.2.1
numpy==1.24.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
phonenumbers==8.13.4
Pillow==9.4.0
polib==1.1.1
prices==1.1.1
promise==2.3
prompt-toolkit==2.0.9
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sniffio==1.3.0
social-auth-core==4.3.0
soupsieve==2.4
SQLAlchemy==1.4.46
sqlparse==0.4.3
stringcase==1.2.0
sympy==1.11.1
tablib==3.4.0
text-unidecode==1.3
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
